+++ performances/src/profilers/profilingUtils.ts
@@ -1,4 +1,4 @@
-import type { ProfilingOptions } from './types'
+import type { ProfilingOptions } from '../types'
 
 export function isSdkBundleUrl(options: ProfilingOptions, url: string) {
   return url === options.bundleUrl
