+++ performances/src/profilers/startMemoryProfiling.ts
@@ -1,6 +1,6 @@
 import type { CDPSession } from 'puppeteer'
-import { isSdkBundleUrl } from '../utils'
 import type { ProfilingOptions } from '../types'
+import { isSdkBundleUrl } from './profilingUtils'
 
 export async function startMemoryProfiling(options: ProfilingOptions, client: CDPSession) {
   await client.send('HeapProfiler.enable')
